move notebook intro to docs and replace images
I used screenshots from my own machine so that the images are more
consistent in the styling. I also added a shot of the dropdown menu for
creating a new notebook.

Also removed the `<center>` tags since they aren't necessary when
displayed on the site and seem to mess stuff up below where they are located.